+++ src/components/Button.svelte
@@ -1,0 +1,21 @@
+<script lang="ts">
+	import clsx from 'clsx'
+
+	export let disabled = false
+</script>
+
+<button
+	{disabled}
+	class={clsx(
+		$$props.class,
+		'h-[fit-content] px-2 py-1 bg-gray-300 text-blue-primary rounded-sm transition-all duration-200',
+		'pointer:hover:text-gray-50 pointer:hover:bg-blue-primary',
+		'active:bg-blue-primary active:text-gray-50',
+		disabled &&
+			'pointer-events-none cursor-not-allowed bg-gray-200 text-gray-400',
+		!disabled && 'shadow-all'
+	)}
+	on:click
+>
+	<slot />
+</button>
