+++ src/utils/types/index.ts
@@ -1,0 +1,2 @@
+export type Region = 'East' | 'Midwest' | 'Rockies' | 'West' | ''
+export type Pass = 'Epic' | 'Ikon'
