+++ vite.config.ts
@@ -2,5 +2,8 @@ import { sveltekit } from '@sveltejs/kit/vite'
 import { defineConfig } from 'vite'
 
 export default defineConfig({
+	optimizeDeps: {
+		include: ['@genkit-ai/google-genai']
+	},
 	plugins: [sveltekit()]
 })
